♻️ strips out the width syncing logic to a separate hook
